Refactor initial Data service code.
